Hack away failing dependencies
Do not need for this hack the PhantomJS (which is deprecated anyway)
or sass. Gulp-sass with old version in the package.json had problems
on install (some node-gyp compilation error) so just changed the package
version to new one. As a result the styles do not work in the plugin.

However that is fine for this purpose since we are interested only getting
the raw data out of the plugin as json.